# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "This is the OpenAPI specification for your project."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "CreateQuizAndQuestionsSchema": {
        "type": "object",
        "properties": {
          "quiz": {
            "$ref": "#/components/schemas/CreateQuizSchema"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateQuestionSchema"
            },
            "minItems": 4
          }
        },
        "required": ["quiz", "questions"]
      },
      "QuizCreationSuccessResponseSchema": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "quizId": {
            "type": "string"
          }
        },
        "required": ["success", "quizId"]
      },
      "QuizListQuerySchema": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "number",
            "minimum": 1,
            "maximum": 20,
            "default": 10,
            "description": "Max quizzes returned per query"
          },
          "cursor": {
            "type": "string",
            "description": "Cursor for Pagination"
          },
          "sortBy": {
            "type": "string",
            "enum": ["asc", "desc"],
            "default": "asc",
            "description": "Sort based on creation date"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryEnum"
            },
            "description": "Only return quizzes containing these tags"
          }
        },
        "required": ["limit", "sortBy"]
      },
      "QuizListResponseSchema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicQuizSchema"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["data"]
      },
      "IDSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": ["id"]
      },
      "QuizDetailResponseSchema": {
        "type": "object",
        "properties": {
          "quiz": {
            "type": "object",
            "properties": {
              "questions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PublicQuestionSchema"
                }
              }
            },
            "required": ["questions"]
          }
        },
        "required": ["quiz"]
      },
      "ResponseBase": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": ["success"]
      },
      "UserPublicSchema": {
        "type": "object",
        "properties": {
          "emailVerified": {
            "type": "boolean"
          },
          "image": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["Learner", "Educator"]
          }
        },
        "required": [
          "emailVerified",
          "email",
          "name",
          "createdAt",
          "id",
          "role"
        ]
      },
      "UserPartialSchema": {
        "type": "object",
        "properties": {
          "emailVerified": {
            "type": "boolean"
          },
          "image": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["Learner", "Educator"]
          }
        }
      },
      "QuizQuestionSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "quizId": {
            "type": "string"
          },
          "order": {
            "type": "string",
            "minimum": 0
          },
          "question": {
            "type": "string",
            "minLength": 5,
            "maxLength": 100
          },
          "answers": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 2
            },
            "minItems": 2
          },
          "correctAnswer": {
            "type": "string",
            "minimum": 0
          }
        },
        "required": [
          "id",
          "quizId",
          "order",
          "question",
          "answers",
          "correctAnswer"
        ]
      },
      "CategoryEnum": {
        "type": "string",
        "enum": [
          "Mathematics",
          "Science",
          "History",
          "Language",
          "Geography",
          "Technology",
          "General"
        ]
      },
      "QuizSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "title": {
            "type": "string",
            "minLength": 5
          },
          "description": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryEnum"
          },
          "numQuestions": {
            "type": "string",
            "minimum": 0
          }
        },
        "required": [
          "id",
          "createdBy",
          "createdAt",
          "title",
          "description",
          "category",
          "numQuestions"
        ]
      },
      "APIErrorSchema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "meta": {
            "type": "string"
          }
        },
        "required": ["message"]
      },
      "PublicQuestionSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "quizId": {
            "type": "string"
          },
          "order": {
            "type": "string",
            "minimum": 0
          },
          "question": {
            "type": "string",
            "minLength": 5,
            "maxLength": 100
          },
          "answers": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 2
            },
            "minItems": 2
          }
        },
        "required": ["id", "quizId", "order", "question", "answers"]
      },
      "PublicQuizSchema": {
        "$ref": "#/components/schemas/QuizSchema"
      },
      "CreateQuestionSchema": {
        "type": "object",
        "properties": {
          "order": {
            "type": "string",
            "minimum": 0
          },
          "question": {
            "type": "string",
            "minLength": 5,
            "maxLength": 100
          },
          "answers": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 2
            },
            "minItems": 2
          },
          "correctAnswer": {
            "type": "string",
            "minimum": 0
          }
        },
        "required": ["order", "question", "answers", "correctAnswer"]
      },
      "CreateQuizSchema": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 5
          },
          "description": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryEnum"
          }
        },
        "required": ["title", "description", "category"]
      },
      "UserSchema": {
        "type": "object",
        "properties": {
          "emailVerified": {
            "type": "boolean"
          },
          "image": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["Learner", "Educator"]
          }
        },
        "required": [
          "emailVerified",
          "email",
          "name",
          "updatedAt",
          "createdAt",
          "id",
          "role"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean",
            "default": false,
            "readOnly": true
          },
          "image": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "default": "Generated at runtime"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "default": "Generated at runtime"
          },
          "role": {
            "type": "string"
          }
        },
        "required": ["name", "email", "createdAt", "updatedAt"]
      },
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "token": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "default": "Generated at runtime"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "ipAddress": {
            "type": "string"
          },
          "userAgent": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": ["expiresAt", "token", "createdAt", "updatedAt", "userId"]
      },
      "Account": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "accountId": {
            "type": "string"
          },
          "providerId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "idToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "refreshTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "scope": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "default": "Generated at runtime"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "accountId",
          "providerId",
          "userId",
          "createdAt",
          "updatedAt"
        ]
      },
      "Verification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "identifier": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "default": "Generated at runtime"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "default": "Generated at runtime"
          }
        },
        "required": [
          "identifier",
          "value",
          "expiresAt",
          "createdAt",
          "updatedAt"
        ]
      }
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Authentication required"
                },
                "code": {
                  "type": "string",
                  "example": "401"
                }
              }
            }
          }
        }
      },
      "403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access denied"
                },
                "code": {
                  "type": "string",
                  "example": "403"
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Resource not found"
                },
                "code": {
                  "type": "string",
                  "example": "404"
                }
              }
            }
          }
        }
      },
      "409": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Resource already exists"
                },
                "code": {
                  "type": "string",
                  "example": "409"
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/quiz": {
      "post": {
        "operationId": "post-quiz",
        "summary": "",
        "description": "Creates a new quiz",
        "tags": ["Quiz"],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizAndQuestionsSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizCreationSuccessResponseSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "get": {
        "operationId": "get-quiz",
        "summary": "",
        "description": "Returns a list of quizzes",
        "tags": ["Quiz"],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "number",
              "description": "Max quizzes returned per query"
            },
            "required": false,
            "description": "Max quizzes returned per query"
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string",
              "description": "Cursor for Pagination"
            },
            "required": false,
            "description": "Cursor for Pagination"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "description": "Sort based on creation date"
            },
            "required": false,
            "description": "Sort based on creation date"
          },
          {
            "in": "query",
            "name": "tags",
            "schema": {
              "type": "array",
              "description": "Only return quizzes containing these tags"
            },
            "required": false,
            "description": "Only return quizzes containing these tags"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizListResponseSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/quiz/{id}": {
      "get": {
        "operationId": "get-quiz-{id}",
        "summary": "",
        "description": "Returns a quiz and its questions and possible answers.",
        "tags": ["Quiz"],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDetailResponseSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorSchema"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "operationId": "delete-quiz-{id}",
        "summary": "",
        "description": "Deletes a quiz.",
        "tags": ["Quiz"],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBase"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorSchema"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/user": {
      "get": {
        "operationId": "get-user",
        "summary": "",
        "description": "Retrieves the authenticated user's profile information using Better-Auth session.",
        "tags": ["User"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublicSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorSchema"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "patch": {
        "operationId": "patch-user",
        "summary": "",
        "description": "Updates the current user's information. Only partial data is required.",
        "tags": ["User"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPartialSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublicSchema"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorSchema"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  }
}